moved to sampling design form only necessary fields
