+++ core/i18n/resources/en/common.js
@@ -744,6 +744,7 @@ It can be simple text or Markdown language (https://www.markdownguide.org).`,
         dimensionsSeparate: 'Dimensions separately',
       },
       reportingArea: 'Total reporting area (ha) (Optional)',
+      reportingAreaInfo: `With stratified sampling, give areas of strata within the stratum attribute's category table (column name 'area')`,
     },
     stratumAttribute: 'Stratum attribute',
     postStratificationAttribute: 'Post stratification attribute',
