+++ core/i18n/resources/pt/common.js
@@ -745,6 +745,7 @@ Obrigado e aproveite **$t(common.appNameFull)**!`,
         dimensionsSeparate: 'Dimensões separadamente',
       },
       reportingArea: 'Área total de reporte (ha) (Opcional)',
+      reportingAreaInfo: `Com amostragem estratificada, forneça as áreas dos estratos na tabela de categorias do atributo de estrato (nome da coluna 'area')`,
     },
     stratumAttribute: 'Atributo de estrato',
     postStratificationAttribute: 'Atributo de pós-estratificação',
