+++ webapp/views/App/views/Analysis/Chain/StatisticalAnalysis/ChainStatisticalAnalysis.scss
@@ -1,0 +1,55 @@
+.app-module .chain .statistical-analysis {
+  justify-self: unset;
+
+  .form_row {
+    display: flex;
+    gap: 3rem;
+
+    .form-item {
+      display: flex;
+    }
+
+    &.checkbox-with-info {
+      gap: 0rem;
+      align-items: center;
+    }
+
+    &.p-value {
+      margin-left: 16%;
+    }
+  }
+
+  .entity-selector,
+  .attributes-selector {
+    width: 30rem;
+  }
+
+  .entity-selector-wrapper {
+    display: flex;
+
+    .warning-badge {
+      position: relative;
+    }
+  }
+
+  .dimensions-selector,
+  .filter {
+    width: 40rem;
+  }
+
+  .dropdown-wrapper.p-value-dropdown {
+    width: 7rem;
+
+    .dropdown {
+      width: 7rem;
+    }
+  }
+
+  .form-item.reporting-area {
+    white-space: nowrap;
+
+    input {
+      width: 20rem;
+    }
+  }
+}
