+++ webapp/views/App/views/Analysis/Chain/StatisticalAnalysis/DimensionsSelector/DimensionsSelector.scss
