+++ webapp/views/App/views/Analysis/Chain/StatisticalAnalysis/DimensionsSelector/index.js
